pyproject.toml
[project]
name = "neighbor"
version = "0.1.0"
description = "Neighbor Deep Research Agent - Build evidence-backed, hyper-local profiles of stakeholders around project sites"
readme = "README.md"
authors = [
    { name = "HELPEN Voice", email = "info@helpen.vc" }
]
requires-python = ">=3.10"
dependencies = [
    "openai>=1.68.2",
    "pydantic>=2.0.0",
    "fastapi>=0.115.12",
    "uvicorn[standard]>=0.29.0,<0.30.0",
    "python-dotenv>=1.1.0",
    "jinja2>=3.1.2",
    "playwright>=1.52.0",
    "pypdf>=3.15.0",
    "pyngrok>=7.0.0",
    "aiohttp>=3.9.0",
    "psycopg2-binary>=2.9.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.1.0",
    "pandas>=2.0.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
